[REF] database_synchronization: add test with 'real' self synchronization
