+++ database_synchronization/tests/__init__.py
@@ -1,1 +1,3 @@
 from . import test_module
+from . import test_synchronization_real
+from . import test_synchronization_emulated
